--- __tests__/test_demo.ts
@@ -1,5 +1,0 @@
-// import {sum} from '../src/index';
-
-// test('adds 1 + 2 to equal 3', () => {
-//   expect(sum(1, 2)).toBe(3);
-// });
