+++ src/event_system/index.ts
@@ -1,2 +1,3 @@
-export {AddEventListener, RegistCanvasEvent, RemoveEventListener} from './event_system';
-export {EventType, FssEvent} from './events';
+export {AddEventListener, RegisterCanvasEvent, RemoveEventListener} from './event_system';
+export {EventType} from './event_types';
+export {FssEvent} from './fss_event';
